Suitable problem for testing variance minimization found. Would be interesting to come up with an example where it can't generalize at all, but there is a true pattern there

+++ data_generator.py
@@ -10,5 +10,5 @@ def single_relevant(n, d, magnitude_signal, percent_correct, sd_non_signals):
     signal_1 = np.random.choice([-magnitude_signal, magnitude_signal], size=examples_per_class, p=p)
     signal = np.concatenate([signal_0, signal_1])
     x[:, 0] = signal
-    y = np.concatenate([np.zeros(examples_per_class), np.ones(examples_per_class)])
+    y = np.concatenate([np.zeros(examples_per_class, dtype=int), np.ones(examples_per_class, dtype=int)])
     return x, y
